refactor: move updateVotingProposals to separate service
